conversion proto and robot string
