+++ urdf2webots/importer.py
@@ -47,7 +47,7 @@ def mkdirSafe(directory):
             print('Directory "' + directory + '" already exists!')
 
 
-def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=False,
+def convert2urdf(inFile, outFile=None, isProto=True, robotName='', normal=False, boxCollision=False,
                  disableMeshOptimization=False, enableMultiFile=False,
                  toolSlot=None, initTranslation='0 0 0', initRotation='0 0 1 0',
                  initPos=None, linkToDef=False, jointToDef=False):
@@ -55,8 +55,8 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
         sys.exit('Input file not specified (should be specified with the "--input" argument).')
     if not os.path.isfile(inFile):
         sys.exit('Input file "%s" does not exists.' % inFile)
-    if not inFile.endswith('.urdf') and not inFile.endswith('.PROTO'):
-        sys.exit('"%s" is neither a PROTO file nor an URDF file.' % inFile)
+    if not inFile.endswith('.urdf'):
+        sys.exit('"%s" is not an URDF file.' % inFile)
 
     if not type(initTranslation) == str or len(initTranslation.split()) != 3:
         sys.exit('--translation argument is not valid. Has to be of Type = str and contain 3 values.')
@@ -70,10 +70,7 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
             sys.exit(e, '\n--init-pos argument is not valid. Your list has to be inside of quotation marks. '
                      'Example: --init-pos="1.0, 2, -0.4"')
 
-    if inFile.endswith('.PROTO'):
-        urdf2webots.writeRobot.isPROTO = True
-    else:
-        urdf2webots.writeRobot.isPROTO = False
+    urdf2webots.writeRobot.isProto = isProto
     urdf2webots.parserURDF.disableMeshOptimization = disableMeshOptimization
     urdf2webots.writeRobot.enableMultiFile = enableMultiFile
     urdf2webots.writeRobot.toolSlot = toolSlot
@@ -113,7 +110,7 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
 
         for child in domFile.childNodes:
             if child.localName == 'robot':
-                if urdf2webots.writeRobot.isPROTO:
+                if isProto:
                     if outFile:
                         if os.path.splitext(os.path.basename(outFile))[1] == '.proto':
                             robotName = os.path.splitext(os.path.basename(outFile))[0]
@@ -130,28 +127,21 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
 
                     if enableMultiFile:
                         mkdirSafe(outputFile.replace('.proto', '') + '_meshes')  # make a dir called 'x_meshes'
-                        urdf2webots.writeProto.meshFilesPath = outputFile.replace('.proto', '') + '_meshes'
+                        urdf2webots.writeRobot.meshFilesPath = outputFile.replace('.proto', '') + '_meshes'
+
+                    protoFile = open(outputFile, 'w')
+                    urdf2webots.writeRobot.header(protoFile, inFile, robotName)
                 else:
-                    robotName = convertLUtoUN(urdf2webots.parserURDF.getRobotName(child))  # capitalize
                     tmp_robot_file = tempfile.NamedTemporaryFile(mode="w+", prefix='tempRobotURDFStringWebots')
 
                 urdf2webots.writeRobot.robotName = robotName
                 urdf2webots.parserURDF.robotName = robotName  # pass robotName
 
                 robot = child
-
-
-
-                if urdf2webots.writeRobot.isPROTO:
-                    protoFile = open(outputFile, 'w')
-                    urdf2webots.writeProto.header(protoFile, inFile, robotName)
-
-
-
                 linkElementList = []
                 jointElementList = []
                 for child in robot.childNodes:
-                    elif child.localName == 'link':
+                    if child.localName == 'link':
                         linkElementList.append(child)
                     elif child.localName == 'joint':
                         jointElementList.append(child)
@@ -167,14 +157,18 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
                 childList = []
                 rootLink = urdf2webots.parserURDF.Link()
 
-                for joint in jointElementList:
-                    jointList.append(urdf2webots.parserURDF.getJoint(joint))
-                    parentList.append(jointList[-1].parent)
-                    childList.append(jointList[-1].child)
-                parentList.sort()
-                childList.sort()
                 for link in linkElementList:
                     linkList.append(urdf2webots.parserURDF.getLink(link, inPath))
+                for joint in jointElementList:
+                    jointList.append(urdf2webots.parserURDF.getJoint(joint))
+                if not isProto:
+                    urdf2webots.parserURDF.cleanDummyLinks(linkList, jointList)
+
+                for joint in jointList:
+                    parentList.append(joint.parent)
+                    childList.append(joint.child)
+                parentList.sort()
+                childList.sort()
                 for link in linkList:
                     if urdf2webots.parserURDF.isRootLink(link.name, childList):
                         # We want to skip links between the robot and the static environment.
@@ -194,7 +188,6 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
                                 rootLink = previousRootLink
                                 break
 
-
                         print('Root link: ' + rootLink.name)
                         break
 
@@ -208,12 +201,20 @@ def convert2urdf(inFile, outFile=None, robotName='', normal=False, boxCollision=
                               urdf2webots.parserURDF.Lidar.list)
                 print('There are %d links, %d joints and %d sensors' % (len(linkList), len(jointList), len(sensorList)))
 
-                urdf2webots.writeProto.declaration(protoFile, robotName, initRotation)
-                urdf2webots.writeProto.URDFLink(protoFile, rootLink, 1, parentList, childList, linkList, jointList,
-                                                sensorList, boxCollision=boxCollision, normal=normal, robot=True)
-                protoFile.write('}\n')
-                protoFile.close()
-                return
+                if isProto:
+                    urdf2webots.writeRobot.declaration(protoFile, robotName, initTranslation, initRotation)
+                    urdf2webots.writeRobot.URDFLink(protoFile, rootLink, 1, parentList, childList, linkList, jointList,
+                                                    sensorList, boxCollision=boxCollision, normal=normal, robot=True)
+                    protoFile.write('}\n')
+                    protoFile.close()
+                    return
+                else:
+                    urdf2webots.writeRobot.URDFLink(tmp_robot_file, rootLink, 0, parentList,
+                                childList, linkList, jointList, sensorList, boxCollision=boxCollision,
+                                normal=normal, robot=True, initTranslation=initTranslation, initRotation=initRotation)
+
+                    tmp_robot_file.seek(0)
+                    return (tmp_robot_file.read())
     print('Could not read file')
 
 
@@ -222,6 +223,9 @@ if __name__ == '__main__':
     optParser.add_option('--input', dest='inFile', default='', help='Specifies the urdf file to convert.')
     optParser.add_option('--output', dest='outFile', default='', help='Specifies the path and, if ending in ".proto", name '
                          'of the resulting PROTO file. The filename minus the .proto extension will be the robot name (for PROTO conversion only).')
+    optParser.add_option('--is-proto', dest='isProto', default=True, help='Specifies if the conversion tool will create a PROTO file or '
+                         'output a Webots VRML robot string.')
+    optParser.add_option('--robotName', dest='robotName', default='', help='Specifies the name of the robot (has to be specified if --is-proto is set to False).')
     optParser.add_option('--normal', dest='normal', action='store_true', default=False,
                          help='If set, the normals are exported if present in the URDF definition.')
     optParser.add_option('--box-collision', dest='boxCollision', action='store_true', default=False,
@@ -233,8 +237,10 @@ if __name__ == '__main__':
                          help='If set, the mesh files are exported as separated PROTO files.')
     optParser.add_option('--tool-slot', dest='toolSlot', default=None,
                          help='Specify the link that you want to add a tool slot too (exact link name from urdf, for PROTO conversion only).')
+    optParser.add_option('--translation', dest='initTranslation', default='0 0 0',
+                         help='Set the translation field of your PROTO file or Webots VRML robot string.')
     optParser.add_option('--rotation', dest='initRotation', default='0 0 1 0',
-                         help='Set the rotation field of your PROTO file.')
+                         help='Set the rotation field of your PROTO file or Webots VRML robot string.')
     optParser.add_option('--init-pos', dest='initPos', default=None,
                          help='Set the initial positions of your robot joints. Example: --init-pos="[1.2, 0.5, -1.5]" would '
                          'set the first 3 joints of your robot to the specified values, and leave the rest with their '
@@ -244,5 +250,5 @@ if __name__ == '__main__':
     optParser.add_option('--joint-to-def', dest='jointToDef', action='store_true', default=False,
                          help='If set, urdf joint names are also used as DEF names as well as joint names.')
     options, args = optParser.parse_args()
-    convert2urdf(options.inFile, options.outFile, options.normal, options.boxCollision, options.disableMeshOptimization,
-                 options.enableMultiFile, options.staticBase, options.toolSlot, options.initRotation, options.initPos, options.linkToDef, options.jointToDef)
+    convert2urdf(options.inFile, options.outFile, options.isProto, options.robotName, options.normal, options.boxCollision, options.disableMeshOptimization,
+                 options.enableMultiFile, options.toolSlot, options.initTranslation, options.initRotation, options.initPos, options.linkToDef, options.jointToDef)
